termsDialog: extract component from media3pPane
